factor out quota limiter logic into own module and test suites
